+++ source/B2B.Transactions.CimMessageAdapter/Errors/DuplicateMessageIdDetected.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace B2B.CimMessageAdapter
+namespace B2B.CimMessageAdapter.Errors
 {
     public class DuplicateMessageIdDetected : ValidationError
     {
