+++ source/B2B.Transactions.CimMessageAdapter/Errors/ValidationError.cs
@@ -11,17 +11,17 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-namespace B2B.CimMessageAdapter
+namespace B2B.CimMessageAdapter.Errors
 {
-    public class ValidationError
+    public abstract class ValidationError
     {
-        public ValidationError(string message, string code)
+        protected ValidationError(string message, string code)
         {
             Message = message;
             Code = code;
         }
 
-        public ValidationError(string message)
+        protected ValidationError(string message)
         {
             Message = message;
             Code = "UnknownError";
